+++ tests_maya/conftest.py
@@ -1,4 +1,2 @@
 import pytest
-from tests.fixtures.ktrack_api_fixtures import ktrack_instance, ktrack_instance_non_dropping
 from tests.fixtures.context_fixtures import populated_context
-from tests.fixtures.entity_fixtures import project_dict, shot_dict, asset_dict, task_dict, workfile_dict, user_dict